requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-cors-headers==4.3.1
whitenoise==6.6.0
django-filter==23.5
Pillow==10.2.0
requests==2.31.0
